kitchenService/requirements.txt
# ==========================================
# Production WSGI Server
# ==========================================
gunicorn==21.2.0

# ==========================================
# Django & API Framework
# ==========================================
Django>=5.0
django-filter==24.2
djangorestframework

# ==========================================
# Database & Cache
# ==========================================
django-redis==5.4.0
psycopg[binary]
redis==5.0.1

# ==========================================
# Message Broker
# ==========================================
confluent-kafka==2.3.0

# ==========================================
# API Documentation (Swagger)
# ==========================================
drf-spectacular==0.27.2

# ==========================================
# Utilities & Configuration
# ==========================================
requests
setuptools>=69.0.0